Fix bug and add mit liscence
