+++ src/index.js
@@ -1,6 +1,8 @@
 import './style.css';
-import { addTask, displayTasks, remove } from './functions.js';
-import { complete, clearAllCompleted } from './complete.js';
+import {
+  addTask, displayTasks, remove, clearAllCompleted,
+} from './functions.js';
+import { complete } from './complete.js';
 
 const submit = document.querySelector('.submit');
 submit.addEventListener('click', (e) => {
