+++ src/style.css
@@ -106,7 +106,7 @@ body {
 .make-complete i {
   font-size: 25px;
   color: #808080;
-  cursor: none;
+  
 }
 
 .clear {
